old_stuffs/requirements.txt
flask
line-bot-sdk
python-decouple
python-dotenv
pydash